services/auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.4
aiokafka==0.10.0
httpx==0.27.0
google-auth==2.29.0
requests==2.32.3
python-multipart==0.0.9
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0
python-dotenv==1.0.1
